Only publish reverse proxy ports, configure dataserver/signalmaster forwarding
the containers talk amongst themselves but only the nginx reverse
proxy container publishes a port to the host (currently port 80)

These changes depend on a minimum version of riff-rtc `0.3.0-dev.5`
to work because it added support for configuring the path for the
dataserver websocket connection.

Configuring a path for the signalmaster should also happen,
but that is internal to the SimpleWebRtc package, so it will
take more work. Meantime it is using the default /socket.io path.