extract deployment guide to doc; add python setup, glossary docs, supporting files
The deployment guide is long and complex enough to be better off in a
standalone document.

The python scripts need supporting infrastructure such as packages installed.
Create documentation on setting up a virtual environment w/ the specific
version of python and how to install those packages.

There are a lot of terms when dealing w/ containers and AWS deployment
that can use some explanation. The glossary document is a place to put
those explanations.

Update gitignore so that the python virtual environment files and
byte compiled/optimized/dll files are not committed to the repository.