new structure for docker compose/stack config files
The compose/stack configuration files have been set up to try to avoid having
duplicate configuration where possible.

docker-stack.yml is only intended to be used when deploying images built
using the combined docker-compose.yml and docker-compose.prod.yml

docker-compose.prod.yml is intended to modify the base docker-compose.yml
to create production images.

For the docker-compose.prod.yml:
- The riff-server production Dockerfile is *the* Dockerfile in the root, no
  special prod name.

- The riff-rtc production Dockerfile for running the rtc server is now
  in the docker folder named Dockerfile-prod-server, and that Dockerfile
  depends on the rifflearning/rtc-build image.

- Support specifying (via env var) the tag on the rifflearning/rtc-build
  image that should be used for building the production rtc server and
  web-server.

In Dockerfiles it turns out that only ARGS before the 1st FROM are
usable in any of the FROM directives. Therefore those ARGs are placed
before the comment for the 1st stage instead of after it, I think that
will be a good convention even if it is the only ARG used by a FROM
directive.

Documentation updates describing these files coming soon.