let nginx serve client files, but all routes should be sent to the rtc_upstream
riff-rtc is now a true single page application (SPA) so all routes that
are requested from the server should be forwarded to the rtc_upstream
as long as it needs to dynamically write the response (which it currently
does to add configuration setting).

All other files can be served directly by nginx where they are found either
in the image (production) or on the bind mount volume (development).