move nginx web-server snippets to a snippets directory
Putting the nginx include conf files into a snippets directory seems to be a recommended practice.
Less clear was why some documents still leave a file like proxy_params outside of the
snippets directory, However I found at least one where it was given a .conf extension
and placed in the snippets directory so that is what we're doing here.

https://wiki.debian.org/Nginx/DirectoryStructure
[Docker-based Web Application Setup (example)](https://gist.github.com/koistya/08d3654862f654e756674491593e6fe2)
https://community.letsencrypt.org/t/how-to-work-with-nginxs-snippets-there-are-a-workaround-for-now/35668
https://www.digitalocean.com/community/tutorials/how-to-create-a-self-signed-ssl-certificate-for-nginx-on-debian-8
https://wiki.meurisse.org/wiki/Nginx#snippets