add ssl support into the web-server image
Currently just a self-signed certificate, which will be created in the web-server/ssl
directory if the files don't exist if you run `make build-dev` or `make build-prod`

At this time manual entering of the DN fields is required.

The tls.md doc provides an example of the openssl commands that are run to create the
self-signed cert and the string Diffie-Hellman group and their output.

The conf file is also set to use the `riffplatform.com` domain. For development
I've added it to my local hosts file pointing at 127.0.0.1