Fix Makefile deploy-stack target, add nobuild-vars
Because the `docker stack deploy` does not read in the `.env` file as `docker-compose`
does, but does process variables on keys it doesn't actually use (args), if
FIREBASE_CONFIG is not defined `docker stack deploy` exits with an error.

The nobuild-vars file unsets all of the build environment variables set by build_vars.
This is useful when going back to building dev images where you want to use the
default build variable values.