add env_template intended to be copied to .env and edited
env_template contains commented out settings of all of the docker-compose build
arguments.used at runtime by the created images.
As the values for many of these arguments is secret, they should not be checked in.
Instead copy env_template to .env and enable and set the values as desired before
building images.

This file DOES NOT contain the build variables defined by the `build-vars` file.

Note: It would be worthwhile in the future to look at docker configs and secrets for
passing some of these values to the running containers.