Support deploying to different swarms (beta & staging)
NEEDS DOCUMENTATION written!

you must define an env variable DEPLOY_SWARM with the name of the
swarm (currently beta or staging) this is used to add a final override
compose yml file defining setting specific to the swarm, currently
secret names to the command run by the make target deploy-stack.