beta deploy build-vars and docker-stack.yml changes
We have no process yet for how we record what versions we deploy where and I don't always bother
committing the build-vars (which records the versions I'm building and deploying) or compose file
changes which only change the name of the config or secret which is being used.

When those configs and secrets are checked in somewhere (I'm thinking of committing encrypted files)
then those compose file changes should be checked in along with them.