base docker-compose.yml uses named volumes for redis and mongo data
This is a change where the base docker-compose.yml didn't set those
volumes, the dev yml bound a local directory for the mongo data and
the prod yml used a DIFFERENT named volume.
The redis data was in an anonymous volume.
Although you can delete the volume for the data accidentally a little
more easily than w/ the bound directory, I think it's overall easier
to use.
Naming the volumes for production makes it much easier to keep track
of them. Renaming the volume was needed to mesh w/ changes happening
for the riff mattermost setup.