renewed letsencrypt ssl certs for diagnostic.riffplatform.com
this cert will expire on 2020-04-01

That was the initial cert for diagnostic.
I also removed the certs for sites we no longer are using like beta.riffplatform.com and nsf.riffedu.com
I also unregistered the 2nd account that we don't use so we don't have to choose when renewing any more.