change the image name for the riffrtc and riffdata dev images
There is no need to distinguish the dev image for say riffdata for platform
development from the dev image for edu (mm) development, all that was doing
was tagging the same image twice, making it a little harder to remove when
a new image was built.

The service names are still different w/ a pfm_ prefix for platform and an
edu_ prefix for edu (mm) as having separate containers for development is
useful.