change docker-swarm kill-tunnel behavior
kill-tunnel now only exits w/ non-zero if an existing tunnel was not
killed, while before it exited w/ non-zero if no tunnel was killed because
there was no existing tunnel.

This changes the kill-tunnel script behavior to unset DOCKER_HOST when
no tunnel existed to be killed, which can happen if the tunnel was
started in one shell and killed in another.