remove restart-policy as not applicable to docker in swarm mode
restart-policy doesn't apply to stacks in swarm mode
See https://stackoverflow.com/a/56120617/2184226

For the problem of the services not being restarted when the machine reboots, that seems to be a different issue. I've found this issue [Manager node doesn't rejoin the swarm on reboot](https://github.com/docker/swarmkit/issues/2868) which seems related.