update room to doc mapping
The base document sets are now all in different folders.

the room is no longer the prefix. It needs to be interpreted as a number
where rooms 001-020 map to a prefix 001-020, but rooms 021-040 map to prefix 001-020 and 041-060 also map to prefixes 001-020.

I also attempted to display an error rather than leaving a black screen when there is an issue getting the document from the room name.